LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/MediaStream-MediaElement-preload-none.https-expected.txt

CONSOLE MESSAGE: Error: assert_equals: expected 2 but got 1
When prompted, accept to share your audio and video streams.

This test checks that the HTMLMediaElement preload 'none' attribute value is ignored for MediaStream used as srcObject and MediaStream object URLs used as src.


Harness Error (FAIL), message = Error: assert_equals: expected 2 but got 1

FAIL Test that preload 'none' is ignored for MediaStream object URL used as srcObject for audio assert_unreached: 'suspend' should not be fired. Reached unreachable code
FAIL Test that preload 'none' is ignored for MediaStream used as srcObject for video assert_unreached: 'suspend' should not be fired. Reached unreachable code
